migreate TestTallyDelgatorOverride to simapp
